+++ src/rinex/index.ts
@@ -5,3 +5,8 @@ export * from './warnings';
 export * from './nav';
 export * from './nav-writer';
 export * from './ionex';
+export * from './sp3';
+export * from './obs-writer';
+export * from './obs-writer-v2';
+export * from './obs-writer-v4';
+export * from './obs-writer-common';